requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
sentence-transformers==3.0.1
numpy==2.1.3
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.20
pypdf==5.9.0
PyYAML==6.0.3
reportlab==4.2.5
chardet==5.2.0
